+++ Presentacion/VtnPlanillero.cs
@@ -61,6 +61,24 @@ namespace Presentacion
 
         private void btnGuardar_Click(object sender, EventArgs e)
         {
+            // Verificar si algún campo está vacío
+            if (string.IsNullOrWhiteSpace(txtCedula.Text) ||
+                string.IsNullOrWhiteSpace(txtNombre.Text) ||
+                string.IsNullOrWhiteSpace(txtApellido.Text) ||
+                string.IsNullOrWhiteSpace(txtEmail.Text) ||
+                string.IsNullOrWhiteSpace(txtNac.Text) ||
+                cmbCargos.SelectedItem == null ||
+                string.IsNullOrWhiteSpace(txtContrasena.Text))
+            {
+                MessageBox.Show("Por favor, complete todos los campos antes de guardar.");
+                return; // Sale del método si hay algún campo vacío
+            }
+            // Verificar si el usuario ya existe usando la capa de dominio
+            if (objetoDplanillero.verificarExistenciaUsuario(txtCedula.Text))
+            {
+                MessageBox.Show("Este usuario ya existe.");
+                return; // Sale del método si el usuario ya existe
+            }
             try
             {
                 objetoDplanillero.insertarPlanillero(txtCedula.Text, txtNombre.Text, txtSegundoName.Text, txtApellido.Text, txtSegundoApellido.Text, txtEmail.Text, txtNac.Text, cmbCargos.SelectedItem.ToString(), txtContrasena.Text);
@@ -76,6 +94,13 @@ namespace Presentacion
         }
         private void btnModificar_Click(object sender, EventArgs e)
         {
+            // Verificar si algún campo está vacío
+            if (string.IsNullOrWhiteSpace(txtCedula.Text) ||
+                cmbCargos.SelectedItem == null)
+            {
+                MessageBox.Show("Por favor, complete todos los campos antes de modificar.");
+                return; // Sale del método si hay algún campo vacío
+            }
             try
             {
                 objetoDplanillero.actualizarPlanillero(txtCedula.Text, txtNombre.Text, txtSegundoName.Text, txtApellido.Text, txtSegundoApellido.Text, txtEmail.Text, txtNac.Text, cmbCargos.SelectedItem.ToString(), txtContrasena.Text);
@@ -91,6 +116,12 @@ namespace Presentacion
         }
         private void btnEliminar_Click(object sender, EventArgs e)
         {
+            // Verificar que el campo de cedula no este vacio
+            if (string.IsNullOrWhiteSpace(txtCedula.Text))
+            {
+                MessageBox.Show("Por favor, complete el campo de cedula antes de eliminar. ");
+                return; // Sale del método si hay algún campo vacío
+            }
             try
             {
                 objetoDplanillero.eliminarPlanillero(txtCedula.Text, txtNombre.Text, txtSegundoName.Text, txtApellido.Text, txtSegundoApellido.Text, txtEmail.Text, txtNac.Text, txtContrasena.Text);
@@ -119,6 +150,7 @@ namespace Presentacion
             else
                 MessageBox.Show("Selecione una fila por favor");
         }
+
         private void limpiarCampos()
         {
             txtCedula.Clear();
